tests: move TestSolrResponse into test_search.py
